requirements.txt
<<<<<<< HEAD
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Django==4.2.23
django-tailwind==4.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
fonttools==4.57.0
frozendict==2.4.6
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.1
multitasking==0.0.11
narwhals==1.35.0
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
ta==0.11.0
tqdm==4.67.1
tradingview-ta==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yfinance==0.2.55
PyMuPDF==1.24.12
youtube-transcript-api==0.6.2
pysrt==1.1.2
python-magic==0.4.27
lxml==5.3.0
=======
Django==5.0.2
djangorestframework==3.14.0
nltk==3.8.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pytest==8.0.0
pytest-django==4.8.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.31.0 
>>>>>>> origin/main
